# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant Marketing Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories for the current tenant
      parameters:
        - name: market_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted
...
  /email/generate:
    post:
      operationId: generatePromotionalEmail
      tags: [emailGenerator]
      summary: Generate a promotional email for selected assets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEmailBody"
      responses:
        "200":
          description: Generated email content with coaching tips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateEmailResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /email/saved:
    get:
      operationId: listSavedEmails
      tags: [emailGenerator]
      summary: List saved generated emails for the current tenant
      parameters:
        - name: market_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of saved emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedEmail"
    post:
      operationId: saveGeneratedEmail
      tags: [emailGenerator]
      summary: Save a generated email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveEmailBody"
      responses:
        "201":
          description: Saved email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedEmail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /email/saved/{id}:
    delete:
      operationId: deleteSavedEmail
      tags: [emailGenerator]
      summary: Delete a saved email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Email deleted
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tenantId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - tenantId
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        marketId:
          type: integer
      required:
        - name
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
...
    GenerateEmailBody:
      type: object
      properties:
        assetIds:
          type: array
          items:
            type: integer
          minItems: 1
        platform:
          type: string
          enum: [outlook, hubspot_marketing, hubspot_1to1, dynamics_customer]
        tone:
          type: string
          enum: [professional, friendly, urgent]
        callToAction:
          type: string
        recipientContext:
          type: string
        marketId:
          type: integer
      required:
        - assetIds
        - platform
    GenerateEmailResponse:
      type: object
      properties:
        emailBody:
          type: string
        subjectLineSuggestions:
          type: array
          items:
            type: string
        coachingTips:
          type: array
          items:
            type: string
        platform:
          type: string
        assetTitles:
          type: array
          items:
            type: string
      required:
        - emailBody
        - subjectLineSuggestions
        - coachingTips
        - platform
        - assetTitles
    SaveEmailBody:
      type: object
      properties:
        platform:
          type: string
        emailBody:
          type: string
        subjectLineSuggestions:
          type: array
          items:
            type: string
        coachingTips:
          type: array
          items:
            type: string
        assetTitles:
          type: array
          items:
            type: string
        assetIds:
          type: array
          items:
            type: integer
        tone:
          type: string
        callToAction:
          type: string
        recipientContext:
          type: string
        marketId:
          type: integer
      required:
        - platform
        - emailBody
        - subjectLineSuggestions
        - coachingTips
        - assetTitles
        - assetIds
    SavedEmail:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        platform:
          type: string
        emailBody:
          type: string
        subjectLineSuggestions:
          type: array
          items:
            type: string
        coachingTips:
          type: array
          items:
            type: string
        assetTitles:
          type: array
          items:
            type: string
        assetIds:
          type: array
          items:
            type: integer
        tone:
          type: string
          nullable: true
        callToAction:
          type: string
          nullable: true
        recipientContext:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - platform
        - emailBody
        - subjectLineSuggestions
        - coachingTips
        - assetTitles
        - assetIds
        - createdAt
...